breaks up circleci.go into many files, updates badging, fixes #14)
